+++ examples/variable_arguments.js
@@ -1,12 +1,14 @@
 var typed = require('../typed-function');
 
 // create a typed function with a variable number of arguments
-var sum = typed('...number', function (values) {
-  var sum = 0;
-  for (var i = 0; i < values.length; i++) {
-    sum += values[i];
+var sum = typed({
+  '...number': function (values) {
+    var sum = 0;
+    for (var i = 0; i < values.length; i++) {
+      sum += values[i];
+    }
+    return sum;
   }
-  return sum;
 });
 
 // use the typed function